Move class generation to separate file
