+++ lib/common.js
@@ -1,0 +1,5 @@
+function sanitizeName(name) {
+  return `${name.replace(/[-]/g, '_')}`;
+}
+
+module.exports = { sanitizeName };
